added bfnlibrary

+++ bfn-atlas2/clients/src/main/kotlin/com/bfn/client/dto/InvestorProfileStateDTO.kt
@@ -1,11 +1,12 @@
 package com.bfn.client.dto
 
+import java.math.BigDecimal
 import java.util.*
 
 data class InvestorProfileStateDTO (val issuedBy: String?,
                                     val accountId: String,
-                                    val minimumInvoiceAmount: Double,
-                                    val maximumInvoiceAmount: Double,
-                                    val totalInvestment: Double,
+                                    val minimumInvoiceAmount: BigDecimal,
+                                    val maximumInvoiceAmount: BigDecimal,
+                                    val totalInvestment: BigDecimal,
                                     val defaultDiscount: Double,
                                     var date: Date)
