+++ bfn-atlas2/clients/src/main/kotlin/com/bfn/client/dto/TokenDTO.kt
@@ -1,10 +1,12 @@
 package com.bfn.client.dto
 
+import java.math.BigDecimal
+
 data class TokenDTO(var accountId: String,
                     var invoiceId: String,
                     var tokenIdentifier: String,
-                    var amount: Double,
-                    var invoiceAmount: Double,
+                    var amount: BigDecimal,
+                    var invoiceAmount: BigDecimal,
                     var account: AccountInfoDTO,
                     var issuer: String,
                     var holder: String)
